projetoPaulo/avalicação/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
mysql-connector-python==8.0.26
requests
flask-cors